+++ .gitignore
@@ -1,2 +1,2 @@
-node_modules
-logs/
+node_modules/
+logsJSON/
